CVSL-1127 Only compare formatting when additional data is provided
Also split out toEntity and toModel transformers
