+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/LicenceConditionService.kt
@@ -133,7 +133,7 @@ class LicenceConditionService(
     val updatedConditions = existingConditions.getUpdatedConditions(submittedConditions, removedConditions)
 
     val updatedLicence = licenceEntity.copy(
-      additionalConditions = (newConditions + updatedConditions).onEach { checkFormattedText(it) },
+      additionalConditions = (newConditions + updatedConditions),
       dateLastUpdated = LocalDateTime.now(),
       updatedByUsername = username,
     )
