+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/ToModelTransformers.kt
@@ -1,9 +1,7 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service
 
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.LicenceSummary
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.request.CreateLicenceRequest
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.ProbationSearchResult
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.LicenceStatus
 import java.util.Base64
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.AdditionalCondition as EntityAdditionalCondition
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.AdditionalConditionData as EntityAdditionalConditionData
@@ -24,7 +22,7 @@ import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.ProbationSear
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.StandardCondition as ModelStandardCondition
 
 /*
-** Functions which transform JPA entity objects into their API model equivalents and vice-versa.
+** Functions which transform JPA entity objects into their API model equivalents.
 ** Mostly pass-thru but some translations, so useful to keep the database objects separate from API objects.
 */
 
@@ -58,43 +56,6 @@ fun transformToLicenceSummary(licence: EntityLicence): LicenceSummary {
 
 fun transformToListOfSummaries(licences: List<EntityLicence>): List<LicenceSummary> {
   return licences.map { licence -> transformToLicenceSummary(licence) }
-}
-
-fun transform(createRequest: CreateLicenceRequest): EntityLicence {
-  return EntityLicence(
-    typeCode = createRequest.typeCode,
-    version = createRequest.version,
-    statusCode = LicenceStatus.IN_PROGRESS,
-    nomsId = createRequest.nomsId,
-    bookingNo = createRequest.bookingNo,
-    bookingId = createRequest.bookingId,
-    crn = createRequest.crn,
-    pnc = createRequest.pnc,
-    cro = createRequest.cro,
-    prisonCode = createRequest.prisonCode,
-    prisonDescription = createRequest.prisonDescription,
-    prisonTelephone = createRequest.prisonTelephone,
-    forename = createRequest.forename,
-    middleNames = createRequest.middleNames,
-    surname = createRequest.surname,
-    dateOfBirth = createRequest.dateOfBirth,
-    conditionalReleaseDate = createRequest.conditionalReleaseDate,
-    actualReleaseDate = createRequest.actualReleaseDate,
-    sentenceStartDate = createRequest.sentenceStartDate,
-    sentenceEndDate = createRequest.sentenceEndDate,
-    licenceStartDate = createRequest.licenceStartDate,
-    licenceExpiryDate = createRequest.licenceExpiryDate,
-    topupSupervisionStartDate = createRequest.topupSupervisionStartDate,
-    topupSupervisionExpiryDate = createRequest.topupSupervisionExpiryDate,
-    probationAreaCode = createRequest.probationAreaCode,
-    probationAreaDescription = createRequest.probationAreaDescription,
-    probationPduCode = createRequest.probationPduCode,
-    probationPduDescription = createRequest.probationPduDescription,
-    probationLauCode = createRequest.probationLauCode,
-    probationLauDescription = createRequest.probationLauDescription,
-    probationTeamCode = createRequest.probationTeamCode,
-    probationTeamDescription = createRequest.probationTeamDescription,
-  )
 }
 
 fun transform(licence: EntityLicence): ModelLicence {
@@ -160,21 +121,9 @@ fun transform(licence: EntityLicence): ModelLicence {
   )
 }
 
-// Transform a list of model standard conditions to a list of entity StandardConditions, setting the licenceId
-fun List<ModelStandardCondition>.transformToEntityStandard(licence: EntityLicence, conditionType: String): List<EntityStandardCondition> = map { term -> transform(term, licence, conditionType) }
-
-fun transform(model: ModelStandardCondition, licence: EntityLicence, conditionType: String): EntityStandardCondition {
-  return EntityStandardCondition(
-    licence = licence,
-    conditionCode = model.code,
-    conditionSequence = model.sequence,
-    conditionText = model.text,
-    conditionType = conditionType,
-  )
-}
-
 // Transform a list of entity standard conditions to model standard conditions
-fun List<EntityStandardCondition>.transformToModelStandard(conditionType: String): List<ModelStandardCondition> = filter { condition -> condition.conditionType == conditionType }.map(::transform)
+fun List<EntityStandardCondition>.transformToModelStandard(conditionType: String): List<ModelStandardCondition> =
+  filter { condition -> condition.conditionType == conditionType }.map(::transform)
 
 fun transform(entity: EntityStandardCondition): ModelStandardCondition {
   return ModelStandardCondition(
@@ -185,23 +134,9 @@ fun transform(entity: EntityStandardCondition): ModelStandardCondition {
   )
 }
 
-fun transform(model: ModelAdditionalCondition, licence: EntityLicence, conditionType: String): EntityAdditionalCondition {
-  return EntityAdditionalCondition(
-    conditionVersion = licence.version!!,
-    conditionCode = model.code,
-    conditionCategory = model.category,
-    conditionSequence = model.sequence,
-    conditionText = model.text,
-    conditionType = conditionType,
-    licence = licence,
-  )
-}
-
 // Transform a list of entity additional conditions to model additional conditions
-fun List<EntityAdditionalCondition>.transformToModelAdditional(conditionType: String): List<ModelAdditionalCondition> = filter { condition -> condition.conditionType == conditionType }.map(::transform)
-
-// Transform a list of model additional conditions to entity additional conditions
-fun List<ModelAdditionalCondition>.transformToEntityAdditional(licence: EntityLicence, conditionType: String): List<EntityAdditionalCondition> = map { transform(it, licence, conditionType) }
+fun List<EntityAdditionalCondition>.transformToModelAdditional(conditionType: String): List<ModelAdditionalCondition> =
+  filter { condition -> condition.conditionType == conditionType }.map(::transform)
 
 fun transform(entity: EntityAdditionalCondition): ModelAdditionalCondition {
   return ModelAdditionalCondition(
@@ -218,7 +153,8 @@ fun transform(entity: EntityAdditionalCondition): ModelAdditionalCondition {
 }
 
 // Transform a list of entity additional condition data to model additional condition data
-fun List<EntityAdditionalConditionData>.transformToModelAdditionalData(): List<ModelAdditionalConditionData> = map(::transform)
+fun List<EntityAdditionalConditionData>.transformToModelAdditionalData(): List<ModelAdditionalConditionData> =
+  map(::transform)
 
 fun transform(entity: EntityAdditionalConditionData): ModelAdditionalConditionData {
   return ModelAdditionalConditionData(
@@ -226,18 +162,6 @@ fun transform(entity: EntityAdditionalConditionData): ModelAdditionalConditionDa
     sequence = entity.dataSequence,
     field = entity.dataField,
     value = entity.dataValue,
-  )
-}
-
-// Transform a list of model additional condition data to entity additional condition data
-fun List<ModelAdditionalConditionData>.transformToEntityAdditionalData(additionalCondition: EntityAdditionalCondition): List<EntityAdditionalConditionData> = map { transform(it, additionalCondition) }
-
-fun transform(model: ModelAdditionalConditionData, additionalCondition: EntityAdditionalCondition): EntityAdditionalConditionData {
-  return EntityAdditionalConditionData(
-    dataSequence = model.sequence,
-    dataField = model.field,
-    dataValue = model.value,
-    additionalCondition = additionalCondition,
   )
 }
 
@@ -253,7 +177,8 @@ fun transform(entity: EntityBespokeCondition): ModelBespokeCondition {
 }
 
 // Transform a list of entity additional condition uploads to model additional condition uploads
-fun List<EntityAdditionalConditionUploadSummary>.transformToModelAdditionalConditionUploadSummary(): List<ModelAdditionalConditionUploadSummary> = map(::transform)
+fun List<EntityAdditionalConditionUploadSummary>.transformToModelAdditionalConditionUploadSummary(): List<ModelAdditionalConditionUploadSummary> =
+  map(::transform)
 
 fun transform(entity: EntityAdditionalConditionUploadSummary): ModelAdditionalConditionUploadSummary {
   return ModelAdditionalConditionUploadSummary(
@@ -282,18 +207,6 @@ private fun transform(entity: EntityAuditEvent): ModelAuditEvent {
     eventType = entity.eventType,
     summary = entity.summary,
     detail = entity.detail,
-  )
-}
-
-fun transform(model: ModelAuditEvent): EntityAuditEvent {
-  return EntityAuditEvent(
-    licenceId = model.licenceId,
-    eventTime = model.eventTime,
-    username = model.username,
-    fullName = model.fullName,
-    eventType = model.eventType,
-    summary = model.summary,
-    detail = model.detail,
   )
 }
